refactor: domain 구조 전반적으로 리팩토링
