+++ src/domain/Lotto.js
@@ -2,15 +2,10 @@ import validateLottoNumber from '../validations/validate/LottoNumberValidate.js'
 
 class Lotto {
   #numbers;
-  #ranking;
 
   constructor(numbers) {
     this.#validate(numbers);
     this.#numbers = numbers.sort((a, b) => a - b);
-  }
-
-  set ranking(ranking) {
-    this.#ranking = ranking;
   }
 
   #validate(numbers) {
@@ -23,10 +18,6 @@ class Lotto {
 
   get numbers() {
     return [...this.#numbers];
-  }
-
-  get ranking() {
-    return this.#ranking;
   }
 }
